Improved Sensor class.
- Now .handle() call required.
- Now uses esp32's tasks to read the sensor every <sense interval>.
Fixed sense interval in config set to long instead of int.

+++ ESP32/src/Configuration.h
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  **/
-const long dhtSenseInterval = 2000;
+const int dhtSenseInterval = 2000;
 const int led2NumberOfLeds = 60;
 const int led2CurrentLimit = 1500;
 const int led2Brightness = 80;
